+++ plugin.php
@@ -1,7 +1,7 @@
 <?php
 $id = 'sragpegasushelper';
-$version = '1.0.0';
+$version = '1.0.3';
 $ilias_min_version = '5.2.0';
-$ilias_max_version = '5.3.999';
-$responsible = 'Martin Studer, Nicolas Märchy, Nicolas Schäfli';
+$ilias_max_version = '5.4.999';
+$responsible = 'Martin Studer, Nicolas Märchy, Nicolas Schäfli, Marc Schneiter, Jonathan Gerber';
 $responsible_mail = 'info@studer-raimann.ch';