+++ testing/includefile.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * including this file gives access to all testing-related functionality
+ */
+
+include_once "auxiliaries/auxiliaries.php";
+include_once "auxiliaries/cli.php";
+include_once "auxiliaries/database.php";
+include_once "sources/info.php";
+include_once "sources/target.php";
+include_once "tests/tests.php";
+include_once "tests/suite.php";