+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2]
+### Added
+- testing for REST- and PegasusHelper-Plugins
+- support for ILIAS 5.4
+
+## [1.0.1]
+### Fixed
+- White page while opening ILIAS website with an auth token 
+
 ## [1.0.0]
 ### Added
 - Changelog